+++ src/a-star.js
@@ -1,139 +1,12 @@
 // reference: Chapter 3 of Russell, S. J., Norvig, P., & Chang, M.-W. (2021). Artificial Intelligence: A modern approach. Pearson.
 
+const { State, Node, COLORS, getOutputGrid, printGrid, printState } = require('./state');
 const { manhattanDistance, getSafeMoves } = require('./safe-moves');
 const v8 = require('v8');
 
 const structuredClone = obj => {
   return v8.deserialize(v8.serialize(obj));
 };
-
-class Board {
-  width;
-  height;
-  food;
-  // hazards;
-  snakes;
-  constructor(gameBoard) {
-    this.width = gameBoard.width;
-    this.height = gameBoard.height;
-    this.food = gameBoard.food;
-    // this.hazards = gameBoard.hazards;
-    this.snakes = gameBoard.snakes;
-  }
-}
-
-class State {
-  board;
-  you;
-  constructor(gameState) {
-    this.board = new Board(gameState.board);
-    this.you = gameState.you;
-  }
-}
-
-class Node {
-  state;
-  parent;
-  action;
-  pathCost;
-  constructor(state, parent, action, pathCost) {
-    this.state = state;
-    this.parent = parent;
-    this.action = action;
-    this.pathCost = pathCost;
-  }
-}
-
-const COLORS = {
-  Reset: "\x1b[0m",
-  Bright: "\x1b[1m",
-  Dim: "\x1b[2m",
-  Underscore: "\x1b[4m",
-  Blink: "\x1b[5m",
-  Reverse: "\x1b[7m",
-  Hidden: "\x1b[8m",
-  FgBlack: "\x1b[30m",
-  FgRed: "\x1b[31m",
-  FgGreen: "\x1b[32m",
-  FgYellow: "\x1b[33m",
-  FgBlue: "\x1b[34m",
-  FgMagenta: "\x1b[35m",
-  FgCyan: "\x1b[36m",
-  FgWhite: "\x1b[37m",
-  BgBlack: "\x1b[40m",
-  BgRed: "\x1b[41m",
-  BgGreen: "\x1b[42m",
-  BgYellow: "\x1b[43m",
-  BgBlue: "\x1b[44m",
-  BgMagenta: "\x1b[45m",
-  BgCyan: "\x1b[46m",
-  BgWhite: "\x1b[47m"
-};
-
-function boxChar(color) {
-  return `${color}${String.fromCharCode(9632)}${COLORS.Reset}`;
-}
-
-function getOutputGrid({ board, you }) {
-  const output = [];
-  // print grid
-  for (let row = 0; row < board.height; ++row) {
-    output.push([]);
-    for (let col = 0; col < board.width; ++col) {
-      output[row].push(boxChar(COLORS.FgBlack));
-    }
-  }
-  // print food
-  for (const foo of board.food) {
-    output[board.height-1-foo.y][foo.x] = boxChar(COLORS.FgMagenta);
-  }
-  // print other snakes
-  for (const snake of board.snakes) {
-    if (snake.id != you.id) {
-      for (const segment of snake.body) {
-        try {
-          output[board.height-1-segment.y][segment.x] = boxChar(COLORS.FgRed);
-        } catch (err) {
-          console.error(err);
-        }
-      }
-      try {
-        output[board.height-1-snake.head.y][snake.head.x] = boxChar(COLORS.FgYellow);
-      } catch (err) {
-        console.error(err);
-      }
-    }
-  }
-  // print you
-  for (const segment of you.body) {
-    try {
-      output[board.height-1-segment.y][segment.x] = boxChar(COLORS.FgGreen);
-    } catch (err) {
-      console.error(err);
-    }
-  }
-  try {
-    output[board.height-1-you.head.y][you.head.x] = boxChar(COLORS.FgWhite);
-  } catch (err) {
-    console.error(err);
-  }
-  return output;
-}
-
-function printGrid(grid) {
-  console.error('-----------------------');
-  for (const row of grid) {
-    for (const text of row) {
-      process.stderr.write(`${text} `);
-    }
-    process.stderr.write('\n');
-  }
-  process.stderr.write('\n');
-}
-
-function printState(state) {
-  printGrid(getOutputGrid(state));
-}
 
 function printSearchPath(node) {
   let initial = node;
@@ -379,7 +252,9 @@ class MinHeap {
 
 function expand(node) {
   const s = node.state;
-  return getActions(s).map((action) => {
+  const actions = getActions(s);
+  if (!node.parent) console.error(actions);
+  return actions.map((action) => {
     let newS;
     try {
       newS = getResult(s, action);
@@ -453,7 +328,7 @@ function tailHeuristicCost(node) {
 function aStarSearch(gameState) {
   let isGoal;
   let heuristicCost;
-  if (gameState.you.health < 70) {
+  if (gameState.you.health < 90) {
     console.error('hungry');
     isGoal = isFoodGoal;
     heuristicCost = setupFoodHeuristicCost();
@@ -462,7 +337,7 @@ function aStarSearch(gameState) {
     isGoal = isTailGoal;
     heuristicCost = tailHeuristicCost;
   }
-  const aboutToTimeout = getTimeout(100);
+  const aboutToTimeout = getTimeout(400);
   const goal = bestFirstSearch(
     new State(gameState),
     isGoal,
@@ -487,6 +362,7 @@ function defaultMove(gameState) {
   // try to move forward, otherwise random
   let move;
   const actions = getActions(new State(gameState));
+  console.error(actions);
   console.error('trying to move forward...');
   if (actions.includes('up') && isMovingUp(gameState.you)) move = 'up';
   else if (actions.includes('down') && isMovingDown(gameState.you)) move = 'down';
@@ -502,7 +378,6 @@ function defaultMove(gameState) {
 module.exports = {
   State,
   Node,
-  // printState,
   MinHeap,
   bestFirstSearch,
   aStarSearch,
