+++ src/logic.js
@@ -1,7 +1,7 @@
 const { defaultMove } = require('./a-star');
 
 function info() {
-  console.log("INFO")
+  console.error("INFO")
   const response = {
     apiversion: "1",
     author: "",
@@ -13,11 +13,11 @@ function info() {
 }
 
 function start(gameState) {
-  console.log(`${gameState.game.id} START`)
+  console.error(`${gameState.game.id} START`)
 }
 
 function end(gameState) {
-  console.log(`${gameState.game.id} END\n`)
+  console.error(`${gameState.game.id} END\n`)
 }
 
 function move(pool, gameState) {
@@ -32,7 +32,7 @@ function move(pool, gameState) {
         console.error(err);
         response = defaultMove(gameState);
       }
-      console.log(`${gameState.game.id} MOVE ${gameState.turn}: ${response.move}`);
+      console.error(`${gameState.game.id} MOVE ${gameState.turn}: ${response.move}`);
       resolve(response);
     });
   });
